+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v0.6
+- Algorithm::getBuilder: get current SolutionBuilder.
 - (BREAKING) Two types of neighborhoods, normal and lazy. Lazy generated movements under demand, while eager returns a collection of movements.
 - New event type: ErrorEvent, when an unhandled exception is propagated inside an Executor.
 - Telegram bot, if enabled, sends message onErrorEvent.